checkpoint: pull shared logic for registry and resolvers into shared handler lib
